+++ cmd/gitserver/main.go
@@ -28,6 +28,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/authz"
 	"github.com/sourcegraph/sourcegraph/internal/codeintel/dependencies"
+	"github.com/sourcegraph/sourcegraph/internal/codeintel/dependencies/shim"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/conf/conftypes"
 	"github.com/sourcegraph/sourcegraph/internal/database"
@@ -103,7 +104,7 @@ func main() {
 	db := database.NewDB(sqlDB)
 
 	repoStore := database.Repos(db)
-	depsSvc := dependencies.GetService(db, dependencies.ErrorSyncer)
+	depsSvc := dependencies.GetService(db, shim.NewGitService(db), dependencies.ErrorSyncer)
 	externalServiceStore := database.ExternalServices(db)
 
 	err = keyring.Init(ctx)
