+++ internal/codeintel/dependencies/iface.go
@@ -4,6 +4,7 @@ import (
 	"context"
 
 	"github.com/sourcegraph/sourcegraph/internal/api"
+	"github.com/sourcegraph/sourcegraph/internal/codeintel/dependencies/internal/lockfiles"
 	"github.com/sourcegraph/sourcegraph/internal/codeintel/dependencies/internal/store"
 	"github.com/sourcegraph/sourcegraph/internal/conf/reposource"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
@@ -17,6 +18,8 @@ type Store interface {
 type LockfilesService interface {
 	ListDependencies(ctx context.Context, repo api.RepoName, rev string) ([]reposource.PackageDependency, error)
 }
+
+type GitService = lockfiles.GitService
 
 type Syncer interface {
 	// Sync will lazily sync the repos that have been inserted into the database but have not yet been
