+++ internal/repos/sources.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/sourcegraph/sourcegraph/internal/codeintel/dependencies"
 	"github.com/sourcegraph/sourcegraph/internal/database"
-	"github.com/sourcegraph/sourcegraph/internal/database/dbutil"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/auth"
 	"github.com/sourcegraph/sourcegraph/internal/httpcli"
@@ -98,9 +97,7 @@ type DBSource interface {
 
 // WithDB returns a decorator used in NewSourcer that calls SetDB on Sources that
 // can be upgraded to it.
-func WithDB(db dbutil.DB) func(Source) Source {
-	depsSvc := dependencies.GetService(database.NewDB(db), dependencies.ErrorSyncer)
-
+func WithDB(depsSvc *dependencies.Service) func(Source) Source {
 	return func(src Source) Source {
 		if s, ok := src.(DBSource); ok {
 			s.SetDependenciesService(depsSvc)
